requirements.txt
streamlit
fastapi
uvicorn
requests
pydantic
python-dotenv
google-genai
pdfplumber
python-docx
transformers
torch
langdetect
peft 